Fix strumming release bug - temporary solution
Stop playing strummed keys when starting a new chord or changing chord shape, to avoid cases where a note is never released.
Not fully satisfactory, need to also restart playing the strummed strings immediately, and maybe even not release the notes that are common to both the old and new chord's strummed strings.